refactor(plugin/antd): [Header] update header and make it reusable
