+++ plugins/antd/src/components/layout/HeaderBarLeftSideMenuLayout/Header/Header.tsx
@@ -1,20 +1,17 @@
-import { coreActions, CoreStore, Logo } from '@onr/core';
+import { coreActions, CoreStore, Logo, PageProps } from '@onr/core';
 import { Layout } from 'antd';
 import Link from 'next/link';
-import { ComponentType, FC, ReactNode } from 'react';
 import { FiBarChart } from 'react-icons/fi';
 import { useDispatch, useSelector } from 'react-redux';
 import { UserMenu } from '../../components';
 import { DashHeader } from './styles';
 
-interface Props {
-  HeaderMainSection?: ComponentType;
-  logo?: ReactNode;
-  avatar?: string;
+interface Props extends PageProps {
+  showLogoForDsk?: boolean;
 }
 
-export const Header: FC<Props> = props => {
-  const { HeaderMainSection, avatar, logo } = props;
+export const Header: React.FC<Props> = props => {
+  const { HeaderMainSection, avatar, logo, showLogoForDsk = true } = props;
   const { mobile } = useSelector((store: CoreStore) => store.coreStore);
   const dispatch = useDispatch();
 
@@ -22,26 +19,35 @@ export const Header: FC<Props> = props => {
     <DashHeader>
       <Layout.Header className="header">
         {mobile && (
-          <div className="flex-none">
-            <div onClick={() => dispatch(coreActions.setMobileDrawer())} className="trigger">
-              <FiBarChart size={20} strokeWidth={1} />
+          <div className="relative flex-grow">
+            {logo && (
+              <Logo mobile={mobile}>
+                <Link href="/">{logo}</Link>
+              </Logo>
+            )}
+            <div className="absolute top-0 left-0 flex items-center h-full">
+              <div onClick={() => dispatch(coreActions.setMobileDrawer())} className="trigger">
+                <FiBarChart size={20} strokeWidth={1} />
+              </div>
             </div>
           </div>
         )}
 
-        {logo && (
-          <div className="flex-none">
-            <Logo mobile={mobile}>
-              <Link href="/">{logo}</Link>
-            </Logo>
-          </div>
+        {!mobile && (
+          <>
+            <div className="flex-none">
+              {logo && showLogoForDsk && (
+                <Logo mobile={mobile}>
+                  <Link href="/">{logo}</Link>
+                </Logo>
+              )}
+            </div>
+            <div className="flex-grow">{HeaderMainSection && <HeaderMainSection />}</div>
+            <div className="flex-none">
+              <UserMenu avatar={avatar} />
+            </div>
+          </>
         )}
-
-        <div className="flex-grow">{HeaderMainSection && <HeaderMainSection />}</div>
-
-        <div className="flex-none">
-          <UserMenu avatar={avatar} />
-        </div>
       </Layout.Header>
     </DashHeader>
   );
